Exercise 2
Write a Java program to perform the following tasks:
1. Take 5 integers as user inputs and store them in variables. Calculate and display the
average.
2. If the value of the integer part of average is x, display the value of x + xx / x (e.g. If
the average is 34.576, display the value of 34 + 3434 / 34).
A sample output of the program is given below.
Enter an integer number: 19
Enter an integer number: 11
Enter an integer number: 7
Enter an integer number: 5
Enter an integer number: 7
The average is 9.8
The value is 20