Exercise 6
Q1: Suppose you are developing a Java application for a company. Write a Java class to represent an ‘Employee’
1. Employee has six properties: employee number, first name, last name, designation,
gender, and monthly salary.
2. Employee has one method: displayDetails() which prints out the employee number,
first name, last name, designation and monthly salary.

Q2: Write another Java class with the name ‘Test’ that only has a main method.
1. Inside the main method, create an Employee object named ‘employee1’. Set the
values of the properties of employee1 to the following.
• Employee number: 10908
• First name: Bruce
• Last name: Wayne
• Designation: CEO
• Salary: 12,175, 362.10

2. Call the displayDetails method of employee1 to display the employee1’s details. A
sample output is given below.
The details of the employee 10908.
Full name: Bruce Wayne
Designation: CEO
Salary: 12,175,362.10

3. Inside the main method, create another Employee object named ‘employee2’. Set the
values of the properties of employee2 to the following.
• Employee number: 11548
• First name: Lois
• Last name: Lane
• Designation: Office Administrator
• Salary: 10,815.70

4. Call the displayDetails method of emplyee2 to display the employee2’s details. A
sample output is given below.
The details of the employee 11548.
Full name: Lois Lane
Designation: Office Administrator
Salary: 10,815.70

Q3: Write a Java class to represent a ‘Square’.
1. Square has two properties: length of a side and colour.
2. Square has three methods.
a. displayColour() to display the colour.
b. displayArea() to calculate and display the area.
c. displayPerimeter() to calculate and display the perimeter.
Q3: Use the ‘Test’ class that you created in Q2 for the following tasks.
1. Inside the main method, create a Square object named ‘square1’. Set the values of the
properties of square1 to the following.
• Length of a side: 45
• Colour: “Blue”
2. Inside the main method, create a Square object named ‘square2’. Set the values of the
properties of square2 to the following.
• Length of a side: 31
• Colour: “Red”
3. Call the displayColour, displayArea, displayPerimeter methods of each Square to
display the respective details