Exercise 7
Q1: Write a Java class named ‘Circle’ to represent a circle.

1. Circle class has two properties: radius and colour. Both of them should be private.

2. Circle class has six methods: Two getter methods (getRadius and getColour), two setter methods (setRadius and setColour), getPerimeter, and getArea. All these methods should be public.

a. getRadius method returns the radius value of the circle. setRadius method takes in a value and assigns it to radius property of the circle. getColour and setColour methods do the same for the colour property.

b. getPerimeter method calculates the perimeter of the circle (using 2πr) and
returns it.

c. getArea method calculates the area of the circle (using πr2) and returns it.
d. Use Java Math.PI constant for π.
Q2: Write another Java class with the name ‘TestCircle’ that only has a main method.
1. Inside the main method, create a new Circle named ‘circle1’. circle1 should have the
following property values. Use the setter methods to assign values to the properties.

• radius: 10.1
• colour: Brown

2. Call the getPerimeter and getArea methods to get the perimeter and the area of the circle. Display the retrieved values. A sample output is given below.

The perimeter of circle1 is 63.46
The area of circle1 is 320.47

3. Inside the main method, create another new Circle named ‘circle2’. circle2 should
have the following property values. Use the setter methods to assign values to the properties.

• radius: 15.58
• colour: Pink

4. Call the getPerimeter and getArea methods to get the perimeter and the area of the
circle. Display the retrieved values. A sample output is given below.
The perimeter of circle1 is 97.89
The area of circle1 is 762.58