Exercise 9
Q1: Write a Java class named ‘Test’ that only has a main method.

1. Declare an integer array of size 5 and initialize it with values.

a. Using a for loop, increment the values of all the elements by 1.
b. Using a for loop display the values of the array in the reverse order.
c. Using an enhanced for loop, find the minimum, maximum, total and average of all the elements.

2. Declare a String array of size 6 and initialize it with values.

a. Using a for loop, remove any leading and/or trailing blank spaces in the Strings.
b. Using an enhanced for loop, concatenate all the elements (from first element
to last element, in that order) and store it in a String variable.
c. Split the string in half and store the two parts in two String variables. Check
whether the two Strings are equal.
d. Convert the first String to uppercase and the second String to lowercase.

Q2: Write a Java class named ‘Circle’ to represent a circle.

1. Circle class (refer Laboratory Exercise 5).

a. Circle class has one property: radius (double). It should be private.
b. Write getter and setter methods for the property. They should be public.
c. Circle class has a constructor that takes in values to initialize all the instance variables.

Q3: In the same ‘Test’ class:

1. Declare a Circle array that can store 5 Circle objects. Initialize it with values.
2. Increase radiuses of all the circles in the array by 40%.
3. Using an enhanced for loop, calculate the average radius of all the circles in the array.