Exercise 10
Q1: Write a Java class named ‘Car’ to represent a car.

1. Car class has five properties: transmissionType (char), fuelType (char), topSpeed (double), numberOfDoors (int) and serialNumber (String). All these properties are private. The default value of numberOfDoors is 4.

2. Write getter and setter methods for all these properties.

3. Car class has five constructors.

a. A constructor with no parameters that does not set values for any of the properties.
b. A constructor with five parameters to take in values to set all the properties.
c. A constructor with two parameters to take in values to set the properties: transmissionType and fuelType.
d. A constructor with three parameters to take in values to set the properties: transmissionType, fuelType and serialNumber.
e. A constructor with one parameter to take in a value to set the property:serialNumber.

4. Car class has one method named displayDetails which displays the details of a car (all the properties). This method is public.

Q2: Write another Java class with the name ‘Test’ that only has a main method.

1. Inside the main method, create five car objects using each of the five constructors in the Car class. (Use suitable values).
2. Call the display method of each car object to display each of their details.